refactor(tests): consolidate redundant DI guardrail tests
- Merge 3 Infrastructure*RegistrationGuardrailTests into 1 [Theory]
  parameterized by suffix (Repository, Service, Provider)
- Merge DispatchGuardrailTests + DomainEventGuardrailTests into
  DispatcherRegistrationGuardrailTests (4 facts, 1 file)
- Remove duplicate Post_Dogs_ShouldNotReturn404 from
  RouteMappingGuardrailTests (byte-for-byte identical to PostDogs)

Guardrails/: 14 → 11 .cs files, zero test coverage loss.
