+++ tests/CampFitFurDogs.Api.Tests/Guardrails/RouteMappingGuardrailTests.cs
@@ -1,10 +1,8 @@
-using System.Linq;
 using System.Net;
 using System.Net.Http.Json;
 using FluentAssertions;
 
 namespace CampFitFurDogs.Api.Tests.Guardrails;
-
 
 public class RouteMappingGuardrailTests
     : GuardrailTestBase, IClassFixture<CampFitFurDogsApiFactory>
@@ -14,24 +12,6 @@ public class RouteMappingGuardrailTests
 
     [Fact]
     public async Task PostDogs_ShouldNotReturn404()
-    {
-        var client = Factory.CreateClient();
-
-        var request = new
-        {
-            Name = "Test",
-            Breed = "Test",
-            DateOfBirth = "2020-01-01",
-            Sex = "Female"
-        };
-
-        var response = await client.PostAsJsonAsync("/api/dogs", request);
-
-        response.StatusCode.Should().NotBe(HttpStatusCode.NotFound);
-    }
-
-    [Fact]
-    public async Task Post_Dogs_ShouldNotReturn404()
     {
         var client = Factory.CreateClient();
